Fix transliteration with source_lang != "en"
The transliterate function only worked when the source language was set to "en".

The problem was that the transliterate function would turn the source_lang text
into english, and then turn that english part into whatever the destination
language is...

Except that when turning the text from the converted-english part
into the destination language, it never took the converted-english part,
but rather the non-decoded text.